notification rework + some improvements

+++ packages/api/src/subscriptions/services/subscription.service.ts
@@ -1,7 +1,8 @@
 import { BadRequestException, Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { Types } from 'mongoose';
 import { MailSubjects } from 'src/common/constants/enums/mail-subjects.enum';
-import { MailSenderService } from 'src/notifications/mail/mail-sender.service';
+import { NotificationType } from 'src/common/constants/enums/notification-type.enum';
+import { NotificationsService } from 'src/notifications/notifications.service';
 import { WeatherService } from 'src/weather/services/weather.service';
 import { Subscription } from '../../database/schemas/subscription.schema';
 import { CreateSubscriptionDto } from '../dtos/create-subscription.dto';
@@ -18,7 +19,7 @@ export class SubscriptionService {
   constructor(
     @Inject('IServiceSubscriptionRepository')
     private readonly subscriptionRepository: IServiceSubscriptionRepository,
-    private readonly mailSenderService: MailSenderService,
+    private readonly notificationsService: NotificationsService,
     private readonly weatherService: WeatherService,
   ) {}
 
@@ -33,21 +34,27 @@ export class SubscriptionService {
 
     const newSubscription = await this.subscriptionRepository.create(subscribeDto);
 
-    await this.mailSenderService.sendConfirmationEmail({
-      to: newSubscription.email,
-      subject: MailSubjects.SUBSCRIPTION_CONFIRMATION,
-      token: newSubscription._id,
-    });
+    await this.notificationsService.sendConfirmationNotification(
+      {
+        to: newSubscription.email,
+        subject: MailSubjects.SUBSCRIPTION_CONFIRMATION,
+        token: newSubscription._id,
+      },
+      NotificationType.EMAIL,
+    );
   }
 
   async confirm(token: string) {
-    //validation is in dto so can transform safely
+    if (!Types.ObjectId.isValid(token)) throw new NotFoundException('Token Not Found');
+
     const objectId = new Types.ObjectId(token);
     const updated = await this.subscriptionRepository.updateById(objectId, { confirmed: true, expiresAt: null });
     if (!updated) throw new NotFoundException('Token Not Found');
   }
 
   async unsubscribe(token: string) {
+    if (!Types.ObjectId.isValid(token)) throw new NotFoundException('Token Not Found');
+
     const objectId = new Types.ObjectId(token);
     const deleted = await this.subscriptionRepository.deleteById(objectId);
     if (!deleted) throw new NotFoundException('Token Not Found');
